add --skip-default-install option
In this mode, when an application is deployed to strong-pm no
installation or build actions are taken. The commit or tarball are
simply extracted to the location on disk and app is started.

This is to work with the new strong-pack based sl-build behaviour that
is optimized for deployment processes where the build is run in an
environment compatible with the production environment and allows for
even more strict production environments since it does not require a
compiler or even npm to be available at run time.